requirements.txt
aiosmtpd==1.4.6
alembic==1.13.2
atpublic==5.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
rq==1.16.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
